pyproject.toml
[tool.poetry]
name = "ebun"
version = "0.1.0"
description = "This is a bot that generates an 'oddly satisfying' video based on a prompt"
authors = ["Andrew David <andrewdvd10@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
black = "^24.2.0"
python-dotenv = "^1.0.1"
praw = "^7.7.1"
langchain = "^0.1.8"
huggingface-hub = "^0.20.3"
apscheduler = "^3.10.4"
wikipedia-api = "^0.6.0"
redis = "^5.0.3"
langchain-community = "^0.0.29"
google-generativeai = "^0.4.1"
langchain-anthropic = "^0.1.6"
pydantic-settings = "^2.2.1"
boto = "^2.49.0"
sentence-transformers = "^2.6.0" 
torch={path = "/home/emperorsixpacks/Documents/torch-2.2.2-cp311-cp311-manylinux1_x86_64.whl"}
google-cloud-texttospeech = "^2.16.3"
boto3 = "^1.34.90"
tts = "^0.22.0"
ffmpeg-python = "^0.2.0"
redisvl = "^0.2.0"
edgedb = "^1.9.0"
pytest = "^8.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

